Allow for customizing the marker for list items
This isn't something that I expect to be common, but it's part of the
specification. It's important to note that CommonMark does not support
task lists so I left off support for them from this exporter. The
`ox-zola` exporter will handle task lists.

I also left off the tag processing because I wasn't able to figure out
how to get a tag on an item. I see that it's part of the Org Element
spec, but nothing I tried was able to populate the value.