Add support for center blocks
Center blocks do not have a CommonMark feature, but they end up being
simple HTML elements, which are allowed in CommonMark. As such, I added
them to the CommonMark exporter instead of the Zola one.

Two bits of configuration allow you to change the class and whether or
not the inline styles are printed by the exporter.